+++ src/commands/Info/ping.ts
@@ -1,12 +1,22 @@
-import { MessageEmbed, Client } from 'discord.js'
+import KClient from './../../lib/FClient'
+import { Message, MessageEmbed } from 'discord.js'
 
-exports.run = async(client: any, message: any) => {
+exports.run = async(client: KClient, message: Message) => {
     await message.delete()
-    const msg = message.channel.send("Flying")
+    const msg = message.channel.send("Pinging...")
+    const embed = new MessageEmbed()
+    .setAuthor(`${message.author.username}`, `${message.author.displayAvatarURL({ dynamic: true })}`)
+    .setTitle("🏓 Ping Commmand")
+    .addField("🔌 ws/API Latency", `${Math.round(client.ws.ping)}ms`, true)
+    .addField("🗨️ Message Latency", `${(await msg).createdTimestamp-message.createdTimestamp}ms`, true)
+    .setColor("#ffb300")
+    .setTimestamp()
+    ;(await msg).delete()
+    await message.channel.send(embed)
 }
 
-exports.help ={
+exports.help = {
     name:"ping",
     description:"Ping Pong",
-    aliases: ["Yes"]
+    aliases: ["latency"]
 }