+++ src/events/debug.ts
@@ -1,0 +1,3 @@
+module.exports = async (client: any, info:any) => {
+    // console.log(info)
+}