+++ src/events/error.ts
@@ -1,0 +1,3 @@
+module.exports = async (client: any, error:any) => {
+    console.log(error)
+}