+++ src/events/warn.ts
@@ -1,0 +1,3 @@
+module.exports = async (client: any, warn:any) => {
+    console.log(warn)
+}